Sanity test refactored for extension
